organize routes and components
