+++ src/routes/HostPage.js
@@ -1,0 +1,3 @@
+export default function HostPage() {
+  return <div>this is the host page</div>;
+}
