+++ src/routes/JoinPage.js
@@ -1,0 +1,3 @@
+export default function JoinPage() {
+  return <div>this is the join page</div>;
+}
