+++ src/routes/LocalPage.js
@@ -1,0 +1,3 @@
+export default function LocalPage() {
+  return <div>this is the local page</div>;
+}
