+++ src/routes/TriviaTestPage.js
@@ -1,0 +1,7 @@
+export default function TriviaTestPage() {
+  return (
+    <div>
+      <h1>trivia demo</h1>
+    </div>
+  );
+}
